+++ unikmer/cmd/dump.go
@@ -325,7 +325,7 @@ func init() {
 	RootCmd.AddCommand(dumpCmd)
 
 	dumpCmd.Flags().StringP("out-prefix", "o", "-", `out file prefix ("-" for stdout)`)
-	dumpCmd.Flags().BoolP("unique", "u", false, `remove duplicated k-mers`)
+	dumpCmd.Flags().BoolP("unique", "u", false, `remove duplicate k-mers`)
 	dumpCmd.Flags().BoolP("canonical", "K", false, "save the canonical k-mers")
 	dumpCmd.Flags().BoolP("canonical-only", "O", false, "only save the canonical k-mers. This flag overides -K/--canonical")
 	dumpCmd.Flags().BoolP("sorted", "s", false, "input k-mers are sorted")